+++ services/src/northbound-service/nb-api/src/main/java/org/openkilda/northbound/dto/v2/flows/FlowRequestV2.java
@@ -38,7 +38,7 @@ public class FlowRequestV2 {
     private boolean ignoreBandwidth;
     private boolean periodicPings;
     private String description;
-    private Integer maxLatency;
+    private Long maxLatency;
     private Integer priority;
 
     private String diverseFlowId;
